+++ Platipus.Wallet.Api/Application/Responses/Anakatech/Base/AnakatechBaseResponse.cs
@@ -2,7 +2,7 @@
 
 public abstract record AnakatechBaseResponse(
     bool Success,
-    int Balance,
+    long Balance,
     string? ErrorCode,
     double CashBalance = 0,
     double BonusBalance = 0);
