+++ Platipus.Wallet.Api/Application/Results/Anakatech/IAnakatechResult.cs
@@ -5,10 +5,11 @@ using Base;
 public interface IAnakatechResult : IBaseResult<AnakatechErrorCode>
 {
     public bool Success { get; }
-    public int Balance { get; }
-    
+
+    public long Balance { get; }
+
     public double CashBalance { get; }
-    
+
     public double BonusBalance { get; }
 
     public string? ErrorCode { get; }
